requirements.txt
bs4==0.0.2
fastapi==0.111.1
fastapi-cli==0.0.4
hydra-core==1.3.2
langchain==0.2.10
langchain-core==0.2.26
langchain-openai==0.1.17
langchain-groq==0.1.9
langchain-community==0.2.9
lxml==5.1.1
markdownify==0.12.1
pymongo==4.7.3
python-dotenv==1.0.1
qdrant-client[fastembed]>=1.8.2
typer==0.12.3
git+https://github.com/zurk/pretty_logging@a0010b663ee590a73dd75df4e96c307adabf1190
